libpst, lipstzin0

+++ libalign/tests/242_r2_align_enum/test_r2_align_enum.c
@@ -2,7 +2,7 @@
 #define PROG_DESC "test of {r2_align_enum.h}"
 #define PROG_VERS "1.0"
 
-/* Last edited on 2024-12-05 10:20:18 by stolfi */ 
+/* Last edited on 2025-02-16 19:43:37 by stolfi */ 
 /* Created on 2007-07-11 by J. Stolfi, UNICAMP */
 
 #define test_align_COPYRIGHT \
@@ -53,7 +53,7 @@ int32_t main(int32_t argc, char **argv)
   {
     srandom(4615*417);
 
-    for (uint32_t bal = 0;  bal < 2; bal++)
+    for (int32_t bal = 0;  bal < 2; bal++)
       { ralent_do_test(2, (bool_t)bal);
         ralent_do_test(5, (bool_t)bal);
       }
@@ -89,7 +89,7 @@ void ralent_do_test(int32_t ni, bool_t bal)
 
     fprintf(stderr, "... Finding the largest dimension of {\\RF} ...\n");
     double ursup = 0.0;
-    for (uint32_t k = 0;  k < nd; k++)  { if (urad[k] > ursup) { ursup = urad[k]; } }
+    for (int32_t k = 0;  k < nd; k++)  { if (urad[k] > ursup) { ursup = urad[k]; } }
     fprintf(stderr, "largest radius of {\\RF} = %.8f\n", ursup);
 
     fprintf(stderr, "... Choosing the optimum point {popt} in {\\RF} ...\n");
@@ -106,11 +106,11 @@ void ralent_do_test(int32_t ni, bool_t bal)
         while (TRUE)
           { assert(ntry < 1000); /* Should not happen... */
             double b[nd]; /* A random vector in the unit ball. */
-            rn_throw_ball(nd, b);
-            for (uint32_t i = 0;  i < ni; i++) 
-              { for (uint32_t j = 0;  j < 2; j++) 
+            rn_throw_ball((uint32_t)nd, b);
+            for (int32_t i = 0;  i < ni; i++) 
+              { for (int32_t j = 0;  j < 2; j++) 
                   { double dij = 0.0;
-                    for (uint32_t k = 0;  k < nd; k++) 
+                    for (int32_t k = 0;  k < nd; k++) 
                       { r2_t *uk = &(U[k*ni]);
                         dij += b[k]*urad[k]*uk[i].c[j];
                       }
@@ -130,7 +130,7 @@ void ralent_do_test(int32_t ni, bool_t bal)
       }
     else
       { /* Use the center: */
-        for (uint32_t i = 0;  i < ni; i++) { popt[i] = ctr[i]; }
+        for (int32_t i = 0;  i < ni; i++) { popt[i] = ctr[i]; }
       }
 
     /* If {nd} is 2, define the axis vectors of the indep plot variables: */
@@ -154,7 +154,7 @@ void ralent_choose_ctr(int32_t ni, r2_t ctr[])
     double cmax = 1.995;
     fprintf(stderr, "... choosing the center {ctr} ...\n");
     r2_align_throw_ctr(ni, cmax, ctr, TRUE);
-    for (uint32_t i = 0;  i < ni; i++) { ctr[i] = (r2_t){{ 1.0, 2.0 }}; }
+    for (int32_t i = 0;  i < ni; i++) { ctr[i] = (r2_t){{ 1.0, 2.0 }}; }
     r2_align_print_vector(stderr, ni, "ctr", -1, ctr);
     return;
   }
@@ -180,7 +180,7 @@ void ralent_test_align_enum(int32_t ni, r2_t ctr[], r2_t arad[], bool_t bal, dou
         also writes the data points to {wr}. */
       
     r2_t psol[ni];
-    for (uint32_t i = 0;  i < ni; i++) { psol[i] = ctr[i]; }
+    for (int32_t i = 0;  i < ni; i++) { psol[i] = ctr[i]; }
     double F2ini = FD2(ni, psol);
     fprintf(stderr, "  F2 (ini) = %12.6f\n", F2ini);
 
@@ -192,15 +192,15 @@ void ralent_test_align_enum(int32_t ni, r2_t ctr[], r2_t arad[], bool_t bal, dou
     r2_align_enum(ni, &FD2, arad, bal, tol, psol, &F2sol);
     
     fprintf(stderr, "  F2 (sol) = %12.6f\n", F2sol);
-    for (uint32_t i = 0;  i < ni; i++) 
+    for (int32_t i = 0;  i < ni; i++) 
       { fprintf(stderr, "  psol[%d] = (", i);
-        for (uint32_t j = 0;  j < 2; j++) 
+        for (int32_t j = 0;  j < 2; j++) 
           { fprintf(stderr, " %12.6f", psol[i].c[j]); }
         fprintf(stderr, " ) popt = (");
-        for (uint32_t j = 0;  j < 2; j++) 
+        for (int32_t j = 0;  j < 2; j++) 
           { fprintf(stderr, " %12.6f", popt[i].c[j]); }
         fprintf(stderr, " ) diff = (");
-        for (uint32_t j = 0;  j < 2; j++) 
+        for (int32_t j = 0;  j < 2; j++) 
           { fprintf(stderr, " %12.6f", psol[i].c[j] - popt[i].c[j]); }
         fprintf(stderr, " )\n");
       }
@@ -222,8 +222,8 @@ void ralent_test_align_enum(int32_t ni, r2_t ctr[], r2_t arad[], bool_t bal, dou
         double F2val = r2_align_dist_sqr(ni, q, popt);
         if (plot && (wr != NULL))
           { r2_t dsmp[ni];
-            for (uint32_t i = 0;  i < ni; i++) 
-              { for (uint32_t j = 0;  j < 2; j++) 
+            for (int32_t i = 0;  i < ni; i++) 
+              { for (int32_t j = 0;  j < 2; j++) 
                   { dsmp[i].c[j] = q[i].c[j] - ctr[i].c[j]; }
               }
             double s0 = r2_align_dot(ni, dsmp, u0);
